api codeigniter

+++ app/Controllers/ClientesController.php
@@ -26,34 +26,48 @@ class ClientesController extends BaseController
         // Pega os clientes do banco de dados com paginação
         $clientes = $this->clienteModel->orderBy("id", "desc")->paginate($perPage, 'clientes');
 
-        // Passa os clientes e o pager para a view
-        return view('clientes/index', [
+        // Obtém o total de registros
+        $total = $this->clienteModel->countAllResults();
+
+        // Prepara a resposta JSON
+        $response = service('response');
+        $response->setJSON([
             'clientes' => $clientes,
-            'pager' => $this->clienteModel->pager,
-            'currentPage' => $currentPage
+            'paginateInfo' => [
+                'currentPage' => $currentPage,
+                'perPage' => $perPage,
+                'totalRecords' => $total,
+                'totalPages' => ceil($total / $perPage)
+            ]
         ]);
-    }
 
-
-    public function new()
-    {
-        // Retorna o formulário para criar um novo cliente
-        return view('clientes/create');
+        // Retorna a resposta JSON
+        return $response->getBody();
     }
 
     public function create()
     {
         $rules = [
-            'nome' => 'required|min_length[3]|max_length[10]',
+            'nome' => 'required|min_length[3]|max_length[100]',
         ];
 
+        $response = service('response');
+
         if ($this->validate($rules)) {
-            $data = $this->request->getPost();
+            $data = $this->request->getJSON(true); // para ele retornar um array direcional ao invés de um stdClass
             $this->clienteModel->save($data);
-            return redirect()->to('/clientes')->with('success', 'Cliente criado com sucesso.');
+
+            $id = $this->clienteModel->getInsertID();
+            $cliente = $this->clienteModel->find($id);
+
+            $response->setStatusCode(201);
+            $response->setJSON($cliente);
         } else {
-            return redirect()->back()->withInput()->with('errors', $this->validator->getErrors());
+            $response->setStatusCode(400); // HTTP Status Code para 'Bad Request'
+            $response->setJSON(['errors' => $this->validator->getErrors()]);
         }
+
+        return $response->getBody();
     }
 
     public function show($id)
@@ -61,17 +75,13 @@ class ClientesController extends BaseController
         // Pega o cliente específico do banco de dados
         $cliente = $this->clienteModel->find($id);
 
-        // Passa o cliente para a view
-        return view('clientes/show', ['cliente' => $cliente]);
-    }
+        if (!$cliente) {
+            throw new \CodeIgniter\Exceptions\PageNotFoundException("Cliente com ID $id não encontrado.");
+        }
 
-    public function edit($id)
-    {
-        // Pega o cliente específico do banco de dados
-        $cliente = $this->clienteModel->find($id);
-
-        // Retorna o formulário de edição
-        return view('clientes/edit', ['cliente' => $cliente]);
+        $response = service('response');
+        $response->setJSON($cliente);
+        return $response->getBody();
     }
 
     public function update($id)
@@ -80,25 +90,44 @@ class ClientesController extends BaseController
             'nome' => 'required|min_length[3]|max_length[100]',
         ];
 
+
+        $response = service('response');
+
         if ($this->validate($rules)) {
-            $data = $this->request->getPost();
+            $data = $this->request->getJSON(true); // para ele retornar um array direcional ao invés de um stdClass
             $this->clienteModel->update($id, $data);
-            return redirect()->to('/clientes')->with('success', 'Cliente atualizado com sucesso.');
+
+            $cliente = $this->clienteModel->find($id);
+
+            $response->setStatusCode(200);
+            $response->setJSON($cliente);
         } else {
-            return redirect()->back()->withInput()->with('errors', $this->validator->getErrors());
+            $response->setStatusCode(400); // HTTP Status Code para 'Bad Request'
+            $response->setJSON(['errors' => $this->validator->getErrors()]);
         }
+
+        return $response->getBody();
     }
 
 
     public function delete($id)
     {
-        if ($this->clienteModel->delete($id)) {
-            $session = session();
-            $session->setFlashdata('success', 'Cliente apagado com sucesso.');
-        } else {
-            $session->setFlashdata('error', 'Não foi possível apagar o cliente.');
+        $cliente = $this->clienteModel->find($id);
+
+        if (!$cliente) {
+            throw new \CodeIgniter\Exceptions\PageNotFoundException("Cliente com ID $id não encontrado.");
         }
-    
-        return redirect()->to('/clientes');
+
+        $response = service('response');
+
+        if ($this->clienteModel->delete($id)) {
+            $response->setStatusCode(200);
+            $response->setJSON(["mensagem" => "ID $id excluido com sucesso"]);
+        } else {
+            $response->setStatusCode(400);
+            $response->setJSON(['errors' => "O ID $id não pode ser excluído"]);
+        }
+
+        return $response->getBody();
     }
 }
